+++ test/specs/fulfillmentInboundShipment.pending.js
@@ -8,26 +8,30 @@ var expect = chai.expect;
 
 var amazonMws = require('../../lib/amazon-mws')(accessKey, accessSecret);
 
-describe('Finances', function () {
+describe('Fulfillment Inbound Shipment', function () {
 
     before(function () {
         expect(accessKey).to.be.a('string');
         expect(accessSecret).to.be.a('string');
     });
 
-    it('It should get list of Financial Event Groups using ListFinancialEventGroups Action', async function () {
+    it('It should get Inbound Guidance For SKU using GetInboundGuidanceForSKU Action', async function () {
         var options = {
-            'Version': '2015-05-01',
-            'Action': 'ListFinancialEventGroups',
+            'Version': '2010-10-01',
+            'Action': 'GetInboundGuidanceForSKU',
             'SellerId': config.SellerId,
             'MWSAuthToken': config.MWSAuthToken,
-            'FinancialEventGroupStartedAfter': new Date(13, 12, 2016)
+            'MarketplaceId': config.MarketplaceId,
+            'SellerSKUList.Id.1': config.SKU
         };
 
+        console.log('options ', options);
         expect(options.SellerId).to.be.a('string');
         expect(options.MWSAuthToken).to.be.a('string');
+        expect(options.MarketplaceId).to.be.a('string');
+        expect(options['SellerSKUList.Id.1']).to.be.a('string');
 
-        var response = await amazonMws.finances.search(options);
+        var response = await amazonMws.fulfillmentInboundShipment.search(options);
 
         expect(response).to.be.a('object');
         expect(response).to.have.property('FinancialEventGroupList').to.be.a('object');